# smart-backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  handlers.AdminLoginData:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      id:
        example: 00000000-0000-0000-0000-000000000001
        type: string
      role:
        example: super
        type: string
      username:
        example: admin
        type: string
    required:
    - access_token
    - id
    - role
    - username
    type: object
  handlers.AdminLoginRequest:
    properties:
      password:
        example: admin
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  handlers.AdminLoginResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.AdminLoginData'
      message:
        example: Login successful
        type: string
      success:
        example: true
        type: boolean
    required:
    - message
    - success
    type: object
...
  handlers.AuditLogDetailResponse:
    properties:
      data:
        $ref: '#/definitions/models.AdminAuditLog'
      message:
        example: Audit log retrieved successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  handlers.PaginatedAuditLogResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.AdminAuditLog'
        type: array
      message:
        example: Audit logs retrieved successfully
        type: string
      pagination:
        $ref: '#/definitions/handlers.PaginationMeta'
      success:
        example: true
        type: boolean
    type: object
  handlers.PaginationMeta:
    properties:
      current_page:
        example: 1
        type: integer
      last_page:
        example: 1
        type: integer
      per_page:
        example: 100
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  models.AdminAuditLog:
    properties:
      action:
        description: '"create_user", "update_user", "delete_user", "create_admin",
          "delete_admin", "update_contact", etc.'
        type: string
      admin_id:
        description: Who performed the action
        type: string
      admin_name:
        description: Admin username for quick access (denormalized)
        type: string
      created_at:
        type: string
      details:
        description: JSON with request details (what was changed)
        type: string
      error_message:
        description: Error message if failed
        type: string
      id:
        type: string
      ip_address:
        description: Request IP address
        type: string
      resource_id:
        description: UUID or ID of affected resource
        type: string
      resource_type:
        description: '"user", "admin", "contact", etc.'
        type: string
      status:
        description: '"success" or "failed"'
        type: string
      user_agent:
        description: Request user agent
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@ololo-gate.com
    name: API Support
  description: Secure phone-based authentication backend for Ololo Gate management
    system with dual authentication (users & admins) and role-based access control.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Ololo Gate API
  version: "1.0"
paths:
...
  /api/v1/admin/audit-logs:
    get:
      consumes:
      - application/json
      description: Retrieve audit logs of admin actions (super admin only). Returns
        paginated list of all administrative operations.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by admin ID
        in: query
        name: admin_id
        type: string
      - description: Filter by action type
        in: query
        name: action
        type: string
      - description: Filter by resource type
        in: query
        name: resource_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Audit logs retrieved successfully
          schema:
            $ref: '#/definitions/handlers.PaginatedAuditLogResponse'
        "401":
          description: Unauthorized - invalid or missing admin token
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden - super admin access required
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get admin audit logs
      tags:
      - Admin Audit Logs
  /api/v1/admin/audit-logs/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific audit log entry by ID (super admin only)
      parameters:
      - description: Audit log ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Audit log retrieved successfully
          schema:
            $ref: '#/definitions/handlers.AuditLogDetailResponse'
        "401":
          description: Unauthorized - invalid or missing admin token
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "403":
          description: Forbidden - super admin access required
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Audit log not found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get audit log by ID
      tags:
      - Admin Audit Logs
  /api/v1/admin/login:
    post:
      consumes:
      - application/json
      description: Authenticate admin with username and password, returns permanent
        access token (no expiry)
      parameters:
      - description: Admin credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with permanent token
          schema:
            $ref: '#/definitions/handlers.AdminLoginResponse'
        "400":
          description: Invalid request body or missing credentials
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Admin login
      tags:
      - Admin Authentication
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"